Update web data loading

+++ app/services/db_insert.py
@@ -2,6 +2,8 @@ import os
 import logging
 import json
 import datetime
+from time import sleep
+from app.db.schema import Station, StationLocation, Messages
 
 logger = logging.getLogger(__name__)
 
@@ -16,24 +18,51 @@ def clean_json(input_string):
     )
 
 
-def db_insert(base):
+def db_insert(base, delete=False):
     base_path = "/mnt/ssd/database_proccessed"
     files = os.listdir(base_path)
+
     for file_no, file in enumerate(files):
         try:
             with open(base_path + "/" + file, "r", encoding="utf-8") as f:
                 packets_json = json.load(f)
             packets_json = clean_json(packets_json)
-            packets = []
-            for packet_json in packets_json:
-                packets.append(
+            stations = set()
+            locations = []
+            messages = []
+            for i, packet_json in enumerate(packets_json):
+                stations.add(
+                    Station(
+                        station_id=packet_json.get("from"),
+                        ssid=packet_json.get("ssid"),
+                        symbol=packet_json.get("symbol"),
+                    )
+                )
+                stations.add(
+                    Station(station_id=packet_json.get("to"), ssid=None, symbol="@")
+                )
+                if packet_json.get("latitude"):
+                    locations.append(
+                        {
+                            "station": packet_json.get("from"),
+                            "timestamp": (
+                                (
+                                    datetime.datetime.fromtimestamp(
+                                        packet_json.get("timestamp")
+                                    )
+                                )
+                                if packet_json.get("timestamp")
+                                else datetime.datetime.now()
+                            ),
+                            "latitude": packet_json.get("latitude"),
+                            "longitude": packet_json.get("longitude"),
+                        }
+                    )
+                messages.append(
                     {
-                        "callsign": packet_json.get("from"),
-                        "ssid": None,
-                        "destination": packet_json.get("to"),
+                        "src_station": packet_json.get("from"),
+                        "dst_station": packet_json.get("to"),
                         "path": packet_json.get("path"),
-                        "latitude": packet_json.get("latitude"),
-                        "longitude": packet_json.get("longitude"),
                         "timestamp": (
                             (
                                 datetime.datetime.fromtimestamp(
@@ -41,21 +70,28 @@ def db_insert(base):
                                 )
                             )
                             if packet_json.get("timestamp")
-                            else None
+                            else datetime.datetime.now()
                         ),
-                        "symbol": packet_json.get("symbol"),
                         "comment": packet_json.get("comment"),
                         "raw_packet": packet_json,
                     }
                 )
-            base.alchemy_interface.bulk_insert_alchemy_dicts(packets)
+            print("Inserting")
+            # for loc in locations[:15]:
+            #     print(loc)
+
+            base.alchemy_interface.bulk_insert_alchemy_objs(list(stations), Station)
+            base.alchemy_interface.bulk_insert_alchemy_dicts(locations, StationLocation)
+            base.alchemy_interface.bulk_insert_alchemy_dicts(messages, Messages)
         except Exception as e:
             print(f"Failed with error {e}")
             logger.error(f"Couldn't load json {file} with error {e}")
+            sleep(5)
         try:
-            os.remove(base_path + "/" + file)
-            if file_no % 100 == 0:
-                print(f"{file_no}/{len(files)}")
+            if delete:
+                os.remove(base_path + "/" + file)
+                if file_no % 100 == 0:
+                    print(f"{file_no}/{len(files)}")
         except Exception as e:
             print(f"Failed with error {e}")
             logger.error(f"Couldn't delete object {file} with error {e}")
